requirements.txt
pwntools==4.13.1
python_magic==0.4.27
